Some common views were moved to common folder in order to be reusable
